+++ test/test.js
@@ -1,2 +1,3 @@
 import '../src/weekly-notes/test.js';
 import '../src/shared/test.js';
+import '../src/pull-request-quality/test.js';